refac : ver1 model code refactory
