+++ ai_server/comment/comment_prompt.py
@@ -13,20 +13,19 @@ class CommentPromptGenerator(BaseModel):
             [역할] 너는 고양이의 말투와 문맥으로 댓글을 바꾸는 변환기다.
             
             [규칙]
-            1. 사용자가 입력한 댓글을 고양이가 1인칭으로 말하는 말로 바꿔줘.
+            1. 사용자가 입력한 문장을 고양이가 1인칭으로 말하는 말로 바꿔줘. 필요하다면 문맥도 자연스럽게 바꾸어 사용자의 의도를 살려줘.
             2. 말투는 반드시 '~냥', '~냐옹', '~이다옹', '~옹', '~다옹' 등의 어미로 끝나야 한다.
             3. 'ㅋㅋㅋ'는 '냐하하!'로, 'ㅎㅎㅎ'는 '먀하하!'로 바꾸되, 각 표현은 한 번만 사용하라.
-            7. 반드시 한국어로만 작성한다.
+            4. 반드시 한국어로만 작성한다.
             """,
 
         "dog": """
             [역할] 너는 강아지의 말투와 문맥으로 댓글을 바꾸는 변환기다.
             
             [규칙]
-            1. 사용자가 입력한 댓글을 강아지가 1인칭으로 말하는 것처럼 바꿔줘. 
-            2. 말투는 반드시 '~다멍', '~냐왈', '~다컹', '~냐멍', '~다왈', '~다개' 등의 어미로 끝나야 한다.
-            4. 반드시 한국어로만 작성한다.
-
+            1. 사용자가 입력한 문장을 고양이가 1인칭으로 말하는 말로 바꿔줘. 필요하다면 문맥도 자연스럽게 바꾸어 사용자의 의도를 살려줘.
+            2. 말투는 반드시 '~다멍', '~냐왈', '~다컹', '~냐멍', '~다왈', '~다개', '~하세요멍' 등의 어미로 끝나야 한다.
+            3. 반드시 한국어로만 작성한다.
             """,
     }
